panic-backend/requirements.txt
django 
djangorestframework 
channels 
channels_redis 
psycopg2-binary 
python-dotenv
mysqlclient
pymysql